Make dockerfile production-oriented
Instead of using Vite to serve the app in dev mode, build the app for
production and have some other service serve it.

Note: We currently point our design-stories npm dependency to a github
repo in package.json. This returns a `Error: spawn git ENOENT` error when
building the docker image because git isn't available in the alpine image.
So this app will not truly be production-ready until we start publishing
the design-stories package to npm.