[NavBar] align nav bar with MVP, no more links to pages (#140)
Updates the navbar from the demo branch. I wanted to separate this one out because it changes the dev experience: no more links in the nav bar and you have to go through the user flow. 👍 If we want to add an admin nav menu that only devs can see later, in some later sprint we could, but I think it's good for us to go through the user flow experience for now and the flow is very short at least on the shared platform.

## Changes
- remove links from navbar
- sneaks a navigation related change to the breadcrub on institution details to make it go to the authenticated landing page instead of the unauthenticated landing page

## How to test this PR

1. Does this make the navbar closer to MVP?
2. Can I click the breadcrub on the institution details page and make it back to the authenticated landing page?

## Screenshots

![Screenshot 2024-01-04 at 5 20 55 PM](https://github.com/cfpb/sbl-frontend/assets/19983248/d7a3183d-d951-4568-bfa0-8379b46c98d2)