feat(cicd): add `dependabot.yml` for version update config (#472)
This PR adds the
[`dependabot.yml`](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file)
config file, which defines which package types dependabot should look
for update for, and on what schedule.

We noticed that when we added CodeQL and dependabot at the same time
that it threw some errors. Seems like we need to merge CodeQL first
(#468), _then_ this PR.

**Note:** This is our first time setting up dependabot with
`package-ecosystem: "npm"`. We assume it will work very similar to our
Python projects, but we won't know until with give it a try.

**Note:** The `npm` package type also supports `yarn`. See
[`package-ecosystem`](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#package-ecosystem)
docs.

Co-authored-by: Bill Himmelsbach <whimmels@gmail.com>