fix(ufp-link): change button to link (#663)
`UpdateInstitutionProfile` is a convenient way of linking to the Update
your financial institution profile page, but it was structured as a
button instead of a link. Since the link navigates to a page, it should
be a link: we'll want to do this for accessibility and for styling (line
wrapping). Let me know if I'm missing a reason to keep it a button?

## Changes

- Change `UpdateInstitutionProfile` to be a link instead of a button.

## How to test this PR

1. Do the `update your financial institution profile` links on the
following pages still work?

- View institution profile
- Type of financial institution
- Sign and Submit

## Screenshots

Note: the content here from the Type of financial institution page is
from a future PR, but I'm using it here to show the styling issue.

|Before - Button|After - Link (with visited styling)|
|---|---|
|![Before - Button that does not match
figma](https://github.com/cfpb/sbl-frontend/assets/19983248/53d10351-d0bd-45c1-865a-6de6c8520481)|![After
- Link that matches figma (with visited
styling)](https://github.com/cfpb/sbl-frontend/assets/19983248/80df559c-92e4-49f9-b040-f63a7d8093ff)|